+++ src/main.js
@@ -6,25 +6,16 @@ import VueRouter from 'vue-router'
 // Use VueRoute
 Vue.use(VueRouter)
 
-// Import and define components
+// Import main component
 import App from './App'
-import Home from './components/Home'
-import LoginForm from './components/userComponents/LoginForm'
-import SignUpForm from './components/userComponents/SignUpForm'
-import userProfile from './components/userComponents/userProfile'
-import passwordForgotten from './components/userComponents/passwordForgotten'
+
+import routes from './routes.js'
 
 // Create insance of router and define routes
 const router = new VueRouter({
   mode: 'history',
   base: __dirname,
-  routes: [
-    { path: '/', component: Home, name: 'Home' },
-    { path: '/login', component: LoginForm },
-    { path: '/sign-up', component: SignUpForm },
-    { path: '/profile', component: userProfile },
-    { path: '/forgotten-password', component: passwordForgotten }
-  ]
+  routes
 })
 
 // Extend vue to use router obj and mount to app